TT-1526: Add acceptance test using TestSamlView
The acceptance test uses HtmlUnit to emulate the user's browser and
submits an eIDAS AuthnRequest through an HTML form. On receiving the
AuthnRequest, the proxy node displays a page with information about the
request which is used by the acceptance test.

Code for building `AuthnRequest`s has been moved to `AuthnRequestFactory`.
Code for marshalling a SAMLObject into a string of XML is now in
`SamlMarshaller`. Tests for both of these classes have been included.

Wiremock and the MockHub have been removed for now, since we will
be using Stub IDP to accept and generate SAML messages in place of the
Hub.

Also reverting to the internal Artifactory here (required to run tests
on CI Jenkins).

Author: @vixus0