TT-1674: Publish Hub Metadata for Stub IDP
Stub IDP needs to know which signing certificate is used to sign the
AuthnRequests it receives from Proxy Node and which encryption
certificate to use to encrypt for Proxy Node. Stub IDP expects to be
speaking to Hub, so Proxy Node needs to pretend to be the Hub by
publishing Hub metadata.

Authors: @minhngocd @vixus0