BAU: Fix build.gradle for Gradle 4
Running gradle was issuing the following messages about deprecated
behaviours:

```
Gradle now uses separate output directories for each JVM language, but this build assumes a single directory for all classes from a source set. This behaviour has been deprecated and is scheduled to be removed in Gradle 5.0
The setTestClassesDir(File) method has been deprecated and is scheduled to be removed in Gradle 5.0. Please use the setTestClassesDirs(FileCollection) method instead.
The ConfigurableReport.setDestination(Object) method has been deprecated and is scheduled to be removed in Gradle 5.0. Please use the method ConfigurableReport.setDestination(File) instead.
```

This fixes the acceptanceTest task in `build.gradle` to
use the updated behaviour and remove these error messages.

Author: @vixus0