TT-1707 Accept Authn Without SPType
CEF reference does not always send SPType with the authnrequest. This
appears to be by design because SPType can either be declared in the
Authnrequest or service provider meta data.

A quick fix for now is to default to public SPType until we are able to
consume service provider metadata.

Solo: @ejrowley