TT-1706: Replaced TestKeyPair with TestCertificates that reads from files
Since programmatically generating X509 certificates on the fly in Java
is near impossible without importing BouncyCastle, we have had to
generate it manually and add the X509 cert as a test resource.
TestKeyPair was also generating key-pairs for testing, which is
essentially the same thing, so we have replaced it with reading from
files.

Authors: @cobainc0 @minhngocd