EID-3: Stop Proxy Node serving metadata
In order to validate the Hub response signature, we need to get the
Hub's (currently Stub IDP) signing certificate from metadata. If Proxy
Node is serving this metadata it can't fetch it because the app hasn't
started up yet (chicken/egg type situation). Instead, run a
local_metadata_server (Ruby/WebRick) to serve the metadata for the Hub and
Connector Node to consume.

Authors: @vixus0 @cobainc0