EIDAS-6: Separated the tests out to their own separate classes
Made a conscious decision to use the AuthnRequestBuilder rather than
building the subcomponents for the tests. Reas n being: it better tests the
integration of how we are mapping the AuthnRequest elements, and also
there seem to be defaults set in the SamlBuilder. Happy to discuss this and up for
suggestions.

Solo: @minhngocd