EID-428: Fix Connector Node test metadata
The test metadata for the connector was being built with a default
IDPSSODescriptor and an additional encryption cert that wasn't needed.
This meant that when the credentials were being resolved so the proxy
node could encrypt for the connector, it would sometimes get the wrong
credential, resulting in failing tests.

Author: @vixus0