EID-1058: Remove translation from Gateway
The changes in the commit remove the translation of a hub SAML response
to an eIDAS SAML reponse from the gateway service, and instead POST the
message to the translator for translation/signing.

The gateway handles the response from the translator and returns a SAML
form to the user, as it did before.

This commit also removes unused code from the gateway, as well as
ensuring that only code that is actually shared remains in the
proxy-node-shared project.

Co-authored-by: Tom Rosier <tom.rosier@digital.cabinet-office.gov.uk>
Co-authored-by: dbes-gds <daniel.besbrode@digital.cabinet-office.gov.uk>
Co-authored-by: Chris Wynne <christopher.wynne@digital.cabinet-office.gov.uk>