EID-1029: Use public libraries on Travis
Travis doesn't have access to our Artifactory, so it needs to fetch the
Verify (and other) libs from public repos. There was already a
`VERIFY_USE_PUBLIC_BINARIES` flag in `build.gradle` to deal with this.

Also, `saml-test-utils` is no longer an independent library so removing
it from `build.gradle`.