BAU: Simplify running gradle with Docker
Anything to do with gradle can now be done through the Docker wrapper
`docker-gradle`. `pre-commit.sh` has been modified to use the wrapper, ensuring any tests are run in a consistent
way across machines and operating systems.

We don't need a `Dockerfile.test` as the `gradle-compose.yaml` file
takes care of configuring the Gradle Docker image.

Increased the client timeout in the `AppRule`s as response times are
slightly longer running in the container.