EID-1062 Use fixture id in translator response
The tests now post an eIDAS Authn request to the hub before posting a
hub response. This is so the new request ID watcher validation doesn't
blow up.

This commit ensures that the `inResponseTo` properties of the hub
response, and the eidas response from the translator match, and come
from the same source (`eidas_authn_request.xml` in the test resources).