EID-1155: Dockerize softHSM in Proxy Node Translator
Install softHSM in the Translator container so that the proxy node signing key can be stored in and retrieved from softHSM in test environments.
Modify startup and pki generation scripts to generate the signing key files in a format that can be loaded into softHSM.
Load the signing private key into softHSM in the Dockerfile after installing softHSM so that it is available for use during testing.
Include swedenconnect/opensaml-pkcs11 in the Translator build script so that the library can be used to access keys in softHSM.  Integration tests provided to prove that softHSM and swedenconnect/opensaml-pkcs11 are installed and usable in the container (built from translator.softHSM.Dockerfile).
The docker-compose files have been modified to use the new translator.softHSM.Dockerfile, and include additional arguments and environment variables used to load the key(s) into softHSM.  A separate runtime docker image containing the softHSM installation (softHSM.runtime.Dockerfile) has been created in order to optimise the Translator docker build, so that the softHSM installation is not required to run on every build.  This setup will also facilitate the migration to and setup of AWS CloudHSM which will have different dependencies.
The pre-commit.sh script for Proxy Node uses a separate docker-compose file and image to compile and test the code.  This container does not use the individual proxy node component containers to run the tests, meaning that the softHSM installation was missing for pre-commit.  As such a new Dockerfile has been created to build this 'docker-gradle' service (pre-commit.softHSM.Dockerfile) which includes the softHSM installation too.  This means that the softHSM integration tests are still able to run under pre-commit.

In summary, these changes install and test softHSM and swedenconnect/opensaml-pkcs11 in the Translator Docker container, and prove the ability to load our test keys into softHSM.

The use of a key retrieved from softHSM to sign messages in the proxy node is to be implemented in a subsequent task EID-1160.