EID-1136: Create TranslatorService
The gateway often times out trying to communicate with the translator
due to the really short (500ms) connection timeout setting on the
default `HttpClient`. I've added a `TranslatorServiceConfiguration`
config class that takes the `url` and a `JerseyClientConfiguration` for
the client used to talk to the translator, allowing us to increase the
timeout.

I also pulled the bit that handles communication with the translator out
into its own class `TranslatorService` which is unit-tested in
`TranslatorServiceTest`.

The prod and test configs had drifted out of sync, so I had to update the
`Application`s and `AppRule`s to account for the `redisServerUrl` parameter.