BAU: Configuration refactors
We can use the configuration classes as a factory to build the objects
we need, rather than pulling everything out and `new`ing it up in the
`*Application` class. The `TranslationServiceConfiguration` now builds a
`TranslationService`. The `ReplayCheckerConfiguration` builds a
`MessageReplayChecker`.

There was also an issue with the test config in `src/test/resources` and
the "real" config in `src/dist` drifting out of sync, so I've removed
the test config and now all the tests use the real config file.