begin bundling kubernetes manifests to charts
We have updated the startup helper script to push local docker builds
into the minikube k8s instance rather than relying on external public
images.

For this to work we needed to be able to parameterize which image tags
are referenced (in this case an md5sum of the app dir) so we started
breaking out the apps into individual helm charts which allow us to do:

`helm template charts/proxy-node-translator --set image.tag=local.blah'

to generate the kubernetes resources.

Three charts have been created (gateway, translator, stubs)... we may
want to split stubs down futher in the future, but stubs are still using
the public images so it's not important yet.