EID-1182: Remove hub-saml dependency
The proxy node shouldn't have a dependency on `hub-saml`, which is
really meant for code specific to the Verify Hub. The way it does
validation is also not really in line with the current OpenSAML
implementation. For now, I'm pulling in all the relevant `hub-saml` code
pending creation of a separate SAML validation library.

I've also updated the OpenSAML version and `saml-libs` version in this
commit.