explicitly set --shell=bash to load docker-env
the startup script uses `minikube docker-env` to direct the docker
client on the host machine to the docker daemon in the minikube VM in
order to transfer locally built images directly into the kubernetes
environment.

the docker-env command makes use of the SHELL environment variable to
determin the script to output ready to eval.

this means the script fails for users who are running something like
fish as it generates a fish script and attempts to eval it with bash.

to fix this we explicitly set the --shell=bash