Add liveness and readiness probes to containers
Without liveness and readiness probes, pods will not get automatically
destroyed when reporting as unhealthy and as a result the applications
will not heal themselves in scenerios like ordering of when dependent
applications are available during startup.

This adds probes to the /healthcheck endpoints exposed on the management
(mgmt) ports for the apps that support it.