remove dependency on gradle-compose built image
the gradle tests currently require the softHSM build and the translator
node currently builds an image based on an image that is only built from
the gradle-compose file. This is confusing as it means you cannot build
the translator image without first running the unrelated gradle-compose.

To reduce confusion this removes the gradle-compose and softHSM.base
dockerfile file and replaces them with a single docker-based script and
dedicated dockerfile clearly named after the script that consumes it.