use tcp readiness probes against http port
instead of using httpGet probes against the mgmt /healthcheck endpoint
we should use simple tcp checks against the http port so that when
the app stops accepting new connections (such as during a shutdown)
traffic will stop being routed to the application.

it _may_ be possible that some requests are sent to the app when it is
failing it's healthcheck... but in this scenario the application is
about to get killed due to the failing liveness probe anyway so the
simple tcp check is still prefable.