EID-1184: Restructure Helm charts for independent softhsm
I'm pulling the `softhsm` out into its own container so that it doesn't
need to be inserted into the translator service's container. `softhsm`
now runs as a neighbouring container in the `proxy-node-translator`
deployment, with them sharing an `EmptyDir` volume at `/softhsm` so that
the `libpkcs11-proxy.so` shared object file can be used by the
translator service.

The `proxy-node-translator` deployment needs to know about the `softhsm`
docker image tag and the best way I could find of doing that without
substantially changing the way we build stuff was to turn the current charts
into subcharts of a parent chart, and then passing in the image tags as
global values.