EID-1184: Re-generate pki if minikube IP changes
Running `shutdown.sh` will delete your minikube cluster. On some setups,
recreating the minikube VM leads to it being assigned a new IP. This
means the URLs used to generate the local test PKI are no longer valid.
I've added a file to track the IP used to generate the PKI and if that
doesn't match your minikube VM's IP, the PKI will be regenerated.