EID-1184: Overhaul local build system
Keeping track of components and how to build them in `startup.sh` was
starting to get a bit hairy, so I've created a `docker-compose.yml` to
actually define the apps and their Docker build parameters. In
`startup.sh` I've done a bit of YAML/JSON munging (sorry) to use
`docker-compose` to build and tag the images correctly before
propagating those image tags to the Helm chart template.

At least now there's a single source of truth for what needs to be
built. To correctly tag stuff even when the Java `Dockerfile` changes, I
had to symlink the `Dockerfile` in the root directory into the Java
subproject directories. I then got `tar` to dereference symlinks, so
that the `md5sum` calculation takes `Dockerfile` changes into account.