EID-1254: Generate and Sign Connector Metadata
Re-factoring of local/dev metadata generation scripts to create stand-alone connector metadata generation script that can be run from the command-line.
Some groundwork done to sign the metadata with the hsm, this will be completed in a subsequent task.
The stand-alone script currently does not sign the metadata at all, however the local/dev metadata generation and signing continues to work as-is.