eid1206 enable multiple releases of chart
in order to be able to deploy multiple instances of the proxy-node (as
will likely be required for multiple countries or multiple versions
running in parralell) the names of all kubernetes resources need to be
parameterized with {{ .Release.Name }} so they no longer clash.

as part of this renaming we also addressed the issues that many of the
environment variables referenced the FQDNs of the pods (ie ending in
$namespace.svc.cluster.local) which would prevent them from being
deployed into any namespace other than "default".

work to seperate the softhsm container was also (unfortunatly) bundled
into this commit.