EID-1254: Metadata generator Java app
Instead of trying to tie together scripts and tools across multiple
repos to generate and sign the metadata for the stub connector and proxy
node, we can use this one Java app to do it all.

This commit creates a standalone Java app to generate proxy service and
connector node metadata using a X509 certificate and key file. HSM
support still needs work.