EID-1254: Add support for ECDSA keys
The eIDAS cryptographic specification allows the use of both the
RSASSA-PSS and ECDSA algorithms. I've added a check for the user passing
an EC public key and force the algorithm to ECDSA accordingly.

I've also added a global X509KeyInfoGeneratorFactory that inserts the
X509Certificate into the KeyInfo element in signatures and
SSODescriptors.