EID-1323 Implement client for EidasSamlParserService
This moves the EidasSamlParserService to the gateway as this is the only
place it will be used.

It implements the client for it to call the eidas saml parser, as well
as the configuration.

A service is being generated in the `GatewayApplication` class, ready to
be passed into the new resource when we register it on the environment.