EID-1254: Sign using (Soft)HSM
This enables the generator to sign using both EC and RSA keys stored in
an HSM. Currently the interface is quite SoftHSM-centric but this might
change when we use the CloudHSM for example.

I've added some simple tests using `xmlsectool` to verify the metadata
signature, which you can run using `pre-commit.sh`. Watch out because
it'll wipe your SoftHSM keys!