EID-1317 Check response for errors and validate DTO
The `NotBlank` annotations on the DTO are not actually checked until you
use a validator on the object.

This also checks the status code on the response, as well as if the
entity is correct (can be processed).

If any of these things fail they'll raise a custom error message. This
can later be handled with a mapper.