EID-1328 Add generateAuthnRequest to VSPProxy
This adds a method in the VerifyServiceProvider proxy to call the
generateAuthnRequest endpoint of the VSP.

It also adds a method on the VerifyServiceProviderConfiguration to build
the proxy. This is to reduce boilerplate in the apps main Application
class.

It also consolidates all the shared dtos into one package, contracts,
rather than being split over two.

Finally it updates the VSP dtos with the necessary constructors.