EID-1297 Use relayState and destination from session
The relay state will be stored in the session in the authn request. It's
just a piece of information which the saml standard states should be
returned unaltered, so we just include the one initially sent in the
eidas response.

The destination is confusingly not the destination pulled from the
origin eidas authn request, but the destination the eidas response
should be posted to. The EPS is grabbing this from the connector node
metadata and passing it back to the gateway. We're storing it in the
session in the gateway authn request route.