EID-1313 Map request exceptions to html response
This adds mappers so that if there is a client or server error when
calling the ESP or VSP, the error response is logged and mapped to a
html response for the user.

This html page will need looking at and designing.

It moves the ErrorPageView into shared so that both the ESP and VSP
client can use it.