EID-1313 Make VSP exception and mapper more specific
The exception and mapper was being used for both methods on the vsp
client. They need to return different responses however. The
generateAuthn endpoint needs to return an error page to the user,
however the translate endpoint needs to return a json error response to
the translator.

Making the current exception and mapper specific to the authn endpoint
will allow a more specific exception and mapper to be created for the
translate endpoint to handle this case.