EID-1337 Refactor gateway authn request app rule tests
Previously the tests were dealing with parsing saml, which we don't need
to do - gateway doesn't need to know anything about saml. Instead, we
can test the expected content of the response via xpath.

This also pushes the method for building an authn response up to the
base apprule class as it will be useful for the hub response apprule
tests.