EID-1337 Add apprule tests for gateway hub response resource
This tests the happy path for the hub response endpoint by firing a
request at it and ensuring the correct values in the html response.

To set this up, we first create a session by hitting the authn endpoint
and capturing the session cookie. This is sent in the request to the hub
response endpoint.