EID-1307 Log authn request attributes in ESP
Cyber needs us to log the attributes of the authn request after it's
been validated by the ESP.

They need to attributes to be key value pairs in a json log event. We're
using the logstash-console appender which is jsonifying the logs for us.

We can add custom key/value pairs to those logs by adding them to
[logbacks MDC](https://logback.qos.ch/manual/mdc.html).