EID-1381 Integrate Snyk CLI with Travis
Snyk doesn't play very nicely with gradle sub-projects that use
configurations like we do. As such we can't use the Github integration.
We _can_ use the CLI however with Travis to get a similar result.

We can't test individual sub projects, only the gradle configurations.
We configurations to be defined are in one place and source by the test
and monitor scripts to ensure consistency.

The monitor script is run as part of the `after_success` phase of the
Travis build. This will only be run if all tests pass, as well as the
snyk test. If the monitor script fails for some reason the build won't
be marked as failure.

If the test script fails, the build will be marked as a failure. If any
of the configurations show a security issue, the script will fail. At
this point you can fix the issue. If the issue can't be fixed, it can eb
ignored via the snyk UI.