EID-1357 Integration test with embedded redis for sessions
It was observed at one point that redis could potentially have an issue
with some special characters in certifictes. This seems less risky as
they are all base64 encoded by the ESP before being sent to the gateway
and stored in session.

This test fires up an embedded redis server, rather than using a mock,
and checks that the certificate is retrieved from the session correctly.