EID-1310: Handle internal errors - return SAML error message
Return an error SAML message whenever possible. We need to the Translator to generate and sign the SAML message.
Not all exception mappers can return a SAML message instead of the error page as some of the exceptions are thrown before the session has been properly initialised. We'd need to refactor a number of things to make this work - something to do in a future iteration.

The SAML message exception mappers need a bunch of stuff to be initialised. In a future iteration we should think of a nicer way to generate SAML error messages.

Author: @CharlesIC