EID-1412: Handle internal errors - redirect to static error page in frontend
Redirect to a static error page hosted in verify-frontend when a failure SAML message cannot be generated and returned to the connector node that sent the authn request.

For the time being, we will redirect to the current error page (hosted by the proxy node) if no redirect URL has been configured. This is to enable the app run locally. In the future, once we have a local env, we will want to make the redirect URL mandatory and have the startup script host a static local error page to redirect to.

Author: @CharlesIC