EID-1384 Tweak acceptance tests
As part of getting the acceptance tests to run in the release pipeline
it'd be useful if the ones we have ran green. These tests need more work
in general but that's out of scope for this ticket.

The error page we serve from the frontend does not include the http
status code. Also selenium does not currently support access to the
status code of a request, so the error page test has been changed to
just check the content.