EID-1384 Use Gemfile.lock when building acceptance tests
The Dockerfile wasn't copying the lockfile when building the image,
which resulted in some unexpected behaviour. Capybara 3.14 was being
installed which has an issue with selenium-webdriver 3.141.5926 (loads
of warnings in the logs).

This issue was rectified in capybara 3.17.0.

This also updates the lockfile to use the newest version of things, and
installs bundler v2 to install them.