EID-1488 Reduce number of exception mappers
Delete unused exceptions.

Reduce exception mappers to one that returns a SAML response form and a base class with two sub classes that return a redirect to the error page (one for exceptions we throw and a generic one for when things really did go wrong).

Minor change to build.gradle to import the GrizzlyWebTestContainerFactory so that rules can inject @Context into the mappers as required during testin