EID-1497: Set journey ID on all log events
We need to ensure all log events for a single journey have the same unique ID associated with them.
The journey ID is set in the Gateway on receipt of an authn request. It is then propagated by setting headers and reading information from headers, the HTTP session and MDC in filters on every request received in a resource.

Summary of changes:
  - Set new journey ID in Gateway
  - Propagate journeyId through headers
  - Log ingress and egress message for every received request
  - Log additional information for every request

Co-authored-by: Phil Miller <phillip.miller@digital.cabinet-office.gov.uk>