EID-1513 Add CloudHSM KeyHandle to logging
Use the magic of reflection to call the getHandle method on the CaviumKey because AWS only provide the jar buried inside a CloudHSM client that only runs on Linux or Windows and that's probably a better approach than yanking the jar out and making a package for artifactory so the code builds on our machines.

Pass the key through the code so that it can be logged at the right time.

Co-authored by: Karol Gancarz <karol.gancarz@digital.cabinet-office.gov.uk>