EID-1506 Use ProxyNode Journey ID for relay-state when sending requests to hub
Use the ProxyNode Journey ID for the relay-state when sending request to the hub.  This will help us track journeys across microservice boundaries.

Get the ProxyNode Journey ID from the http session when handling authn requests and put it in the relay-state of the saml form that sends the request to the hub.

Update tests accordingly.

Co-authored by: Karol Gancarz <karol.gancarz@digital.cabinet-office.gov.uk>