BAU: Test generating dependency files for GitHub security alerts
We can take advantage of the GitHub security alerts feature which sends automated alerts for vulnerabilities detected in project dependencies. This can replace Snyk and is directly integrated with GiHub and enabled for projects using the dependency graph feature. The dependency graph feature is not compatible with Gradle, however, and requires generating POM files for each project. This is easily done with a Maven plugin for Grade and can be done automatically when building or testing the project locally.

To also ensure that the dependency files stay up-to-date, we can make Travis open a pull request if it detects a stale file on the master branch. We already use the GitHub API for Codacy coverage report and we can use the same credentials to open pull requests.

Author: @CharlesIC