EID-1532 Use config for PID prefix
The eIDAS specification states that a PID should be prefixed with
"{Identifying country nationality code}/{destination country nationality
code}".

This allows a nationality code to be provided via the configuration.
This is a short term solution as it only allows for the proxy node to be
connected to one country.

When we restructure to allow multiple countries this will need
revisiting.