EID-1508 Set requestlog layout to json
Setting the requestLog type to 'logstash-console' on its own didn't
work. In fact, it meant that no request logs at all were shipped. I
think this was due to dropwizard not being able to find the correct
type.

Rather than using the logstash-console type we can set the layout of the
requestLogs to json, as suggested by the docs: https://www.dropwizard.io/1.3.12/docs/manual/core.html

To do this we need to pull in the dropwizard-json-logging module.