Disable Tagging of docker images as latest
There is an issue when signing docker images that results
in any image signed with the tag "latest" invalidating all
previously signed images.  This is an issue as it prevents
rolling back to a previous release.

This commit disables the signing of tag latest, but should
still result in an image being signed.  Deploying "latest"
should still succeed as the image will be signed any way.