EID-1483 Extend stub-connector authn request error scenarios
Adds an extra error scenario to the stub-connector for an invalid
signature. It uses the test private key from the dev-pki repo. This key
does not match the certificate the proxy-node ESP will be using and so
should result in an invalid signature response.

Also updates the homepage and some styling to differentiate between
valid and invalid journeys.