BAU: Fix Stub Connector entity ID mismatch causing invalid signature errors in ESP
Added a config value to the Stub Connector to ensure it uses the same entity ID in its SAML messages as is advertised in its metadata

Also make chart in the build pipeline a proper pre-release so Concourse won't publish the new release before it passes the tests.

Bumped OpenSAML and Dropwizard versions.

Author: @CharlesIC