BAU: Publish Metadata signing certs from K8s Metadata secret
There are no more GSP admins so we need to have a way to extract the metadata certs to send to countries.

The `metadataCACerts` file is mounted from the cluster. It contains a base64 encoded string containing certs in PEM format.
`MetadataCertsPublishingResource` reads the file, decodes it and then extracts the certs and published them on the specified URL.

Author: @CharlesIC