BAU: stop pipeline stalling if `test` fails
Currently, if `test` fails, then the release will remain in the pre-release state. This means that it will not get picked up by the new run of the pipeline. Thus, the next run of package will just republish the old release and thus test will never trigger again.

We discovered this bug in the DCS pipeline, which is very similar to the proxy node's. See https://github.com/alphagov/doc-checking/commit/c32445c95c055545228013904f2ca25bafafd769 for the equivalent change to the DCS pipeline.