Add metrics from istio
This configures Istio to record HTTP request metrics which include the
request path as a label.

The way this works is: we have a `metric`, which is sent to a
`handler` by a `rule`.

The `metric` records which dimensions from the underlying HTTP event
we wish to record.  The requestcountbypath metric in this commit is
based on the builtin requestcount metric in the istio-system
namespace (see `kubectl -n istio-system get metrics`), but with an
additional `request_path` dimension.

The `rule` states which events are sent to the `handler`.  In this
case, `promhttpbypath` is based on istio-system's `promhttp` rule,
only referring to the requestcountbypath metric instead.  The `match:`
condition states that we will send http events which were not
initiated by kube-probe (which presumably is kubernetes's builtin
healthchecking which we want to ignore).

The `handler` states what to do with the event.  Again, this handler
is based on an istio-system handler called `prometheus`, which
collects events in prometheus metrics.  We have customized this to
`prometheusbypath`, which adds a label `request_path`.  It also
changes the metric name to be namespace-specific.

Note that the metric name has to be namespace-specific.  If you try to
write to the same metric from different namespaces, it gets grumpy.