BAU: Optimise and tidy up eidas-saml-parser tests
Moved `metadataClientRule` and `eidasSamlParserAppRule` out of the base class so they would get initialised in the correct order. Make them static ClassRules to improve performance. Initialise them on creation so they can be static.
Make ESP wait for Connector metadata. This ensures the connector metadata is available when the ESP is being initialised as it fetches the encryption cert when the app first starts.

Use RuleChain to initialise the rules in the correct order.
Make things static where possible.
Rename things according to naming conventions.

Author: @CharlesIC