BAU: Only publish a Github release when acceptance tests have passed
We should really be using an S3 bucket to store temporary build artifacts but that's not available in the cluster yet. We can abuse Docker and use a small image (Alpine) to store the artifacts. A Docker image is just a tar file with a file system and other files in it. We can add the release files to the tar file, push the image and then extract the artifacts in the following steps. This is already done in Sandbox and has two benefits:
  - a github release is only published for tested artifacts
  - the pipeline doesn't get stuck when tests fail and the pre-release cannot be moved up the commit tree
This is a temporary solution and will be replaced by S3 when it becomes available.

  - Remove unnecessary caching of Docker images
  - Add timeouts to deploy tasks

Author: @CharlesIC