BAU Remove xml tag check from validation
Countries using CEF reference connector don't include an XML declaration
at the beginning of Authn requests. This causes requests to fail
validation in the gateway.

This changes the validator to not check for the declaration, and instead
does the minimum of checking the decided Base64 is not blank. If the XML
is actually not valid we'll find out later as it works its way through
the system.