BAU Log ingress and egress on translator failure response endpoint
The egress filter cleans up the MDC for us so we don't log things we
shouldn't if the thread is re-used.

We were failing to do this when the gateway asked the translator to
generate an error response.

This caused an old eIDAS request ID to be logged in Splunk if we ever
got an error, which was confusing and caused an P3 incident with Cyber.