BAU Remove CZ integration deployment
The deployment is currently failing due to the CZ connector-node
metadata endpoint returning a 404. This causes the esp container to
crash loop. We're getting constant alerts in the alerts channel which
are noisy and may make genuine issues harder to spot.

Once the endpoint has been fixed, this commit can be reverted.