EID-1687 Improve handling of missing first name attribute in the Translator
There have been instances of the Proxy Node throwing “Error whilst
handling Hub response. No verified current first name present”.

An attribute can be provided to a Member State if it is verified and
valid for the present time, or if it is valid for the present time.

Currently an attribute is only provided in the former case.

1. Relax logic in the Attribute class to allow
    * the from date to be null,
    * allow no from and to date to be valid for the present time
2. Change translator logic to accept attributes in this order:
    * current and verified
    * If no attributes satisfying a), look for current (without
requiring verified)
3. Log when b) happens so we can track how many times this occurs
4. Log when no current with verified or unverified so we can track which
scenarios we have to deal with

The attribute data to be logged is:
	the attribute type: first name, last name etc.
	is verified
	is validFrom not null
	is validTo not null
The attribute value will not be logged.