Initialise prometheus counters correctly
Currently, the prometheus counters aren't getting initialised
correctly.  If you look at `verify_proxy_node_requests_total` in
Grafana's explore view, you see that counters start from 1, when the
first request happens, not 0, when the app starts.  This means that we
will be missing requests from our count of requests (and of successful
requests).

The problem is less pressing for the response counters, because the
first time *any* counter is accessed, they *all* get initialised.

I think this is because the counters are being created in static
initialisers in MetricUtils.  When do static initialisers run?  It's
complicated, but section 12.4 of the standard is the authority [1].
In particular, static initialisers run when the class is initialised,
which can happen when:

 - a static field is accessed
 - an instance of the class is created

This commit fixes that by moving the counters out of MetricUtils and
into the classes that use them.  This means that the static
initialisers will be run during app startup in the
GatewayApplication.run() method, because this is where instances of
the Resource classes are created.  This means that the Counters should
be initialised to 0 at app startup correctly.

This is similar to the common pattern of putting loggers in the
consuming class with static initialisers.

At this point the MetricUtils class is not terribly useful but I
wasn't sure whether we preferred to keep the code DRY or to inline the
METRIC_PREFIX constant to remove the MetricUtils class.

[1]: https://docs.oracle.com/javase/specs/jls/se11/html/jls-12.html#jls-12.4