EID-1762 Move attribute hash logging logic to saml-lib
We log eIDAS attribute hashes in the proxy-node as well as the hub.
These hashes are compared for security in Splunk.

We want to ensure the same logic is used so the hashes will be the same.

The proxy-node receives attributes from the VSP. The VSP extracts those
attributes from an assertion and serializes them before sending in a
response.

This commit removes the custom Assertion and Attribute objects the
proxy-node has and replaces them with the same classes as used in the
VSP (via saml-lib). This way they should de-serialize into the same
objects. It also imports the new logger that is defined in saml-lib.